pyproject.toml
[tool.poetry]
name = "adapter-prompt"
version = "0.1.0"
description = ""
authors = ["Aman Agrawal"]
readme = "README.md"
packages = [{include = "adapter_prompt"}]

[tool.poetry.dependencies]
python = "3.8"
pandas = "^2.0.0"
scikit-learn = "^1.2.2"
torchvision = "^0.15.1"
adapter-transformers = "^3.2.0"
transformers = "^4.27.4"
peft = "^0.2.0"
dataset = "^1.6.0"
evaluate = "^0.4.0"
bitsandbytes = "^0.37.2"
loguru = "^0.6.0"
gputil = "^1.4.0"
pytorch-transformers = "^1.2.0"
quadprog = "^0.1.11"
gpuutils = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
